Began updating the ViewParser

+++ src/engine/src/engine/ViewParser.cpp
@@ -7,74 +7,70 @@
 
 #include "engine/handlers/ViewParser.hpp"
 
-// ----------------------------------------------------------------------------
-
 #include "engine/handlers/BoolOpParser.hpp"
 #include "engine/handlers/FloatOpParser.hpp"
 #include "engine/handlers/StringOpParser.hpp"
 
-// ----------------------------------------------------------------------------
-
 namespace engine {
 namespace handlers {
 
-void ViewParser::add_column(const Poco::JSON::Object& _obj,
-                            containers::DataFrame* _df) {
-  if (!_obj.has("added_")) {
+void ViewParser::add_column(const ViewOp& _cmd,
+                            containers::DataFrame* _df) const {
+  if (!_cmd.get<"added_">()) {
     return;
   }
 
-  const auto added = *JSON::get_object(_obj, "added_");
+  const auto handle = [this, &_cmd, _df](const auto& _json_col) {
+    using Type = std::decay_t<decltype(_json_col)>;
 
-  const auto json_col = *JSON::get_object(added, "col_");
+    const auto& added = *_cmd.get<"added_">();
 
-  const auto name = JSON::get_value<std::string>(added, "name_");
+    const auto& name = added.get<"name_">();
 
-  const auto role = JSON::get_value<std::string>(added, "role_");
+    const auto& role = added.get<"role_">();
 
-  const auto subroles =
-      JSON::array_to_vector<std::string>(JSON::get_array(added, "subroles_"));
+    const auto& subroles = added.get<"subroles_">();
 
-  const auto unit = JSON::get_value<std::string>(added, "unit_");
+    const auto& unit = added.get<"unit_">();
 
-  const auto type = JSON::get_value<std::string>(json_col, "type_");
+    if constexpr (std::is_same<Type,
+                               commands::FloatColumnOrFloatColumnView>()) {
+      const auto column_view =
+          FloatOpParser(categories_, join_keys_encoding_, data_frames_)
+              .parse(_json_col);
 
-  if (type == FLOAT_COLUMN || type == FLOAT_COLUMN_VIEW) {
-    const auto column_view =
-        FloatOpParser(categories_, join_keys_encoding_, data_frames_)
-            .parse(json_col);
+      auto col = column_view.to_column(0, _df->nrows(), true);
 
-    auto col = column_view.to_column(0, _df->nrows(), true);
+      col.set_name(name);
 
-    col.set_name(name);
+      col.set_subroles(subroles);
 
-    col.set_subroles(subroles);
+      col.set_unit(unit);
 
-    col.set_unit(unit);
+      _df->add_float_column(col, role);
+    } else {
+      const auto column_view =
+          StringOpParser(categories_, join_keys_encoding_, data_frames_)
+              .parse(_json_col);
 
-    _df->add_float_column(col, role);
-  }
+      const auto vec = column_view.to_vector(0, _df->nrows(), true);
 
-  if (type == STRING_COLUMN || type == STRING_COLUMN_VIEW) {
-    const auto column_view =
-        StringOpParser(categories_, join_keys_encoding_, data_frames_)
-            .parse(json_col);
+      assert_true(vec);
 
-    const auto vec = column_view.to_vector(0, _df->nrows(), true);
+      if (role == containers::DataFrame::ROLE_CATEGORICAL ||
+          role == containers::DataFrame::ROLE_JOIN_KEY) {
+        add_int_column_to_df(name, role, subroles, unit, *vec, _df);
+      }
 
-    assert_true(vec);
-
-    if (role == containers::DataFrame::ROLE_CATEGORICAL ||
-        role == containers::DataFrame::ROLE_JOIN_KEY) {
-      add_int_column_to_df(name, role, subroles, unit, *vec, _df);
+      if (role == containers::DataFrame::ROLE_UNUSED ||
+          role == containers::DataFrame::ROLE_TEXT ||
+          role == containers::DataFrame::ROLE_UNUSED_STRING) {
+        add_string_column_to_df(name, role, subroles, unit, *vec, _df);
+      }
     }
+  };
 
-    if (role == containers::DataFrame::ROLE_UNUSED ||
-        role == containers::DataFrame::ROLE_TEXT ||
-        role == containers::DataFrame::ROLE_UNUSED_STRING) {
-      add_string_column_to_df(name, role, subroles, unit, *vec, _df);
-    }
-  }
+  std::visit(handle, _cmd.get<"added_">()->get<"col_">());
 }
 
 // ------------------------------------------------------------------------
@@ -84,7 +80,7 @@ void ViewParser::add_int_column_to_df(const std::string& _name,
                                       const std::vector<std::string>& _subroles,
                                       const std::string& _unit,
                                       const std::vector<strings::String>& _vec,
-                                      containers::DataFrame* _df) {
+                                      containers::DataFrame* _df) const {
   const auto encoding = _role == containers::DataFrame::ROLE_CATEGORICAL
                             ? categories_
                             : join_keys_encoding_;
@@ -113,7 +109,8 @@ void ViewParser::add_int_column_to_df(const std::string& _name,
 void ViewParser::add_string_column_to_df(
     const std::string& _name, const std::string& _role,
     const std::vector<std::string>& _subroles, const std::string& _unit,
-    const std::vector<strings::String>& _vec, containers::DataFrame* _df) {
+    const std::vector<strings::String>& _vec,
+    containers::DataFrame* _df) const {
   auto col = containers::Column<strings::String>(_df->pool());
 
   for (size_t i = 0; i < _vec.size(); ++i) {
@@ -131,10 +128,13 @@ void ViewParser::add_string_column_to_df(
 
 // ----------------------------------------------------------------------------
 
-void ViewParser::drop_columns(const Poco::JSON::Object& _obj,
+void ViewParser::drop_columns(const ViewOp& _cmd,
                               containers::DataFrame* _df) const {
-  const auto dropped =
-      JSON::array_to_vector<std::string>(JSON::get_array(_obj, "dropped_"));
+  if (!_cmd.get<"dropped_">()) {
+    return;
+  }
+
+  const auto& dropped = *_cmd.get<"dropped_">();
 
   for (const auto& name : dropped) {
     _df->remove_column(name);
@@ -342,27 +342,34 @@ Poco::JSON::Object ViewParser::get_content(
 
 // ----------------------------------------------------------------------------
 
-containers::DataFrame ViewParser::parse(const Poco::JSON::Object& _obj) {
-  const auto type = JSON::get_value<std::string>(_obj, "type_");
+containers::DataFrame ViewParser::parse(
+    const commands::DataFrameOrView& _cmd) const {
+  const auto handle = [this](const auto& _cmd) -> containers::DataFrame {
+    using Type = std::decay_t<decltype(_cmd)>;
 
-  if (type == "DataFrame") {
-    const auto name = JSON::get_value<std::string>(_obj, "name_");
-    return utils::Getter::get(name, *data_frames_);
-  }
+    if constexpr (std::is_same<Type, DataFrameOp>()) {
+      const auto name = fct::get<"name_">(_cmd);
+      return utils::Getter::get(name, *data_frames_);
+    } else {
+      const auto& base = *fct::get<"base_">(_cmd);
 
-  const auto base = *JSON::get_object(_obj, "base_");
+      auto df = parse(base);
 
-  auto df = parse(base);
+      add_column(_cmd, &df);
 
-  add_column(_obj, &df);
+      drop_columns(_cmd, &df);
 
-  drop_columns(_obj, &df);
+      subselection(_cmd, &df);
 
-  subselection(_obj, &df);
+      // TODO: Handle the build history.
+      /*df.set_build_history(
+          Poco::JSON::Object::Ptr(new Poco::JSON::Object(_obj)));*/
 
-  df.set_build_history(Poco::JSON::Object::Ptr(new Poco::JSON::Object(_obj)));
+      return df;
+    }
+  };
 
-  return df;
+  return std::visit(handle, _cmd.val_);
 }
 
 // ----------------------------------------------------------------------------
@@ -400,60 +407,62 @@ ViewParser::parse_all(const Poco::JSON::Object& _cmd) {
 
 // ----------------------------------------------------------------------------
 
-void ViewParser::subselection(const Poco::JSON::Object& _obj,
+void ViewParser::subselection(const ViewOp& _cmd,
                               containers::DataFrame* _df) const {
-  if (!_obj.has("subselection_")) {
+  if (!_cmd.get<"subselection_">()) {
     return;
   }
 
-  const auto json_col = *JSON::get_object(_obj, "subselection_");
+  const auto handle = [this, &_cmd, _df](const auto& _json_col) {
+    using Type = std::decay_t<decltype(_json_col)>;
 
-  const auto type = JSON::get_value<std::string>(json_col, "type_");
+    if constexpr (std::is_same<Type, commands::BooleanColumnView>()) {
+      const auto column_view =
+          BoolOpParser(categories_, join_keys_encoding_, data_frames_)
+              .parse(_json_col);
 
-  if (type == BOOLEAN_COLUMN_VIEW) {
-    const auto column_view =
-        BoolOpParser(categories_, join_keys_encoding_, data_frames_)
-            .parse(json_col);
+      const auto data_ptr = column_view.to_vector(0, _df->nrows(), true);
 
-    const auto data_ptr = column_view.to_vector(0, _df->nrows(), true);
+      assert_true(data_ptr);
 
-    assert_true(data_ptr);
+      _df->where(*data_ptr);
+    } else {
+      const auto data_ptr =
+          FloatOpParser(categories_, join_keys_encoding_, data_frames_)
+              .parse(_json_col)
+              .to_vector(0, std::nullopt, false);
 
-    _df->where(*data_ptr);
-  } else {
-    const auto data_ptr =
-        FloatOpParser(categories_, join_keys_encoding_, data_frames_)
-            .parse(json_col)
-            .to_vector(0, std::nullopt, false);
+      assert_true(data_ptr);
 
-    assert_true(data_ptr);
+      const auto& key_float = *data_ptr;
 
-    const auto& key_float = *data_ptr;
+      auto key = std::vector<size_t>(key_float.size());
 
-    auto key = std::vector<size_t>(key_float.size());
+      for (size_t i = 0; i < key_float.size(); ++i) {
+        const auto ix_float = key_float[i];
 
-    for (size_t i = 0; i < key_float.size(); ++i) {
-      const auto ix_float = key_float[i];
+        if (ix_float < 0.0) {
+          throw std::runtime_error(
+              "Index on a numerical subselection cannot be "
+              "smaller than zero!");
+        }
 
-      if (ix_float < 0.0) {
-        throw std::runtime_error(
-            "Index on a numerical subselection cannot be "
-            "smaller than zero!");
+        const auto ix = static_cast<size_t>(ix_float);
+
+        if (ix >= _df->nrows()) {
+          throw std::runtime_error(
+              "Index on a numerical subselection out of "
+              "bounds!");
+        }
+
+        key[i] = ix;
       }
 
-      const auto ix = static_cast<size_t>(ix_float);
-
-      if (ix >= _df->nrows()) {
-        throw std::runtime_error(
-            "Index on a numerical subselection out of "
-            "bounds!");
-      }
-
-      key[i] = ix;
+      _df->sort_by_key(key);
     }
+  };
 
-    _df->sort_by_key(key);
-  }
+  return std::visit(handle, *_cmd.get<"subselection_">());
 }
 
 // ----------------------------------------------------------------------------
